put public key testing in its own script
